- Upgraded to Thrift 0.9.1 - Correct bug where the last updated sync number & time were not saved if it had an issue during the first full sync. - Correct bug where notes with multiple resources wouldn't be displayed properly. - Add transaction support to try and speed up DB activities that are intenseve (like indexing).

+++ thrift/async/TEvhttpClientChannel.cpp
@@ -17,10 +17,10 @@
  * under the License.
  */
 
-#include "TEvhttpClientChannel.h"
+#include <thrift/async/TEvhttpClientChannel.h>
 #include <evhttp.h>
-#include "transport/TBufferTransports.h"
-#include <protocol/TProtocolException.h>
+#include <thrift/transport/TBufferTransports.h>
+#include <thrift/protocol/TProtocolException.h>
 
 #include <iostream>
 #include <sstream>
@@ -142,7 +142,7 @@ void TEvhttpClientChannel::finish(struct evhttp_request* req) {
   }
   recvBuf_->resetBuffer(
       EVBUFFER_DATA(req->input_buffer),
-      EVBUFFER_LENGTH(req->input_buffer));
+      static_cast<uint32_t>(EVBUFFER_LENGTH(req->input_buffer)));
   cob_();
   return;
 }
